Add a 'copySourceToOutput' option
The `sources` element in the generated source map was pointing to `/src/main/scss/....`.
This meant that browsers could not find the original files, reducing the usefulness of
the source map.

This patch updates the code to copy the source file to the target directory before
compiling. This means that the generated source map now points to a file that
will be included in the final war.

I'm not convinced that I've got the name of the option right, but I'm fairly sure the
code works.